soc: qcom: glink_ssr: Export notify_for_subystem() function for testing
notify_for_subsystem() is the function which sends do_cleanup messages
to the appropriate subsystems in the case of SSR. Currently, this
function is not exported and thus cannot be called independently of
an actual SSR event. Allowing this function to be called in this way
will make SSR testing more flexible.

Export this function so that it can be called independently of an actual
SSR event.

Change-Id: I2a2a107f1f180ec98e35ce39272562b95780862f
Signed-off-by: Steven Cahail <scahail@codeaurora.org>