soc: qcom: glink: Allow NULL packet private in queue_rx API
The glink_queue_rx_intent API accepts a void * as one of its parameters
for use by the client.  The client can choose to use this parameter as an
opaque value, which is then returned to the client when the intent is
consumed.  It should be valid to allow the client to specify NULL for this
parameter, indicating that the client does not require this facility.  The
implementation of the glink_queue_rx_intent API does not allow NULL to be
specified.

Fix the glink_queue_rx_intent API to allow the client to specify a NULL
private pointer, which is a valid usecase.

Change-Id: I09eb32b727999e6d66404148a40ecde2f895fa7e
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>