soc: qcom: smem: Fix debugfs entry for smem native transport
The smem transport debugfs initialization  expects the transport
to be already registered with glink to avoid creation of required
directory structure twice. Interrupt information is also not present
in the current debugfs entry.

Fix the debugfs entry by initializing it after the smem transport gets
registered with glink core. The interrupt information is also updated.

Change-Id: I93e7326fedc85dfc8bbce5bf886e17ba4fc4e478
Signed-off-by: Atish Kumar Patra <apatra@codeaurora.org>