mtd: msm_qpic_nand: fix erased page detection
Some of the newer nand parts can have bit flips in an erased
page due to the process technology used. In this case, qpic
nand controller is not able to identify that page as an erased
page.

To workaround this issue, the driver will perform a raw page
read and extract the ECC bytes and check whether they are all
0xff's (as would be the case in an erased page) or not. If the
number of zeroes are below a certain threshold then it can be
safely considered an erased page and will not return an error
to the file system.

CRs-Fixed: 831052
Change-Id: I585abaac96b3b140960b1f6bb8c6459508b0df82
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>