scripts: Restore some of the parallism in "all" builds
The use of min/max should not have been switched, and instead of
forcing at least "-j2" for builds, it is enforcing no more than two
jobs in parallel.  This has drastically reduced the parallelism of the
builds.

Change this to use "max" as intended so that we do make use of
available CPUs, using at least two.

Change-Id: Ifad0c98ef0ce56fc6c2368321a6cfd763e81b370
Signed-off-by: David Brown <davidb@codeaurora.org>