cpufreq: lionfish: standalone implementation
This is a new governor that is optimized for smartphones and similar
battery powered devices. It combines traits of the conservative,
ondemand, and interactive governors, and aims to maintain CPU
frequencies inside an intermediate range with optimal performance per
watt.

It jumps straight to an intermediate frequency when overloaded at low
speed to avoid lag. It jumps to full speed when overloaded above the
intermediate frequency. For moderate loads, it votes to raise, lower,
or maintain the frequency each sampling period. When enough votes to
change the frequency accumulate, it changes the frequency. By default,
this governor also brings down the frequency more aggresively than the
conservative governor, aiming to maintain average CPU load between 40%
and 80%.