mmc: core: fix system resume for cmdq
There is no action taken if the card is runtime suspended by the time
system suspend is triggered. During system resume, the card remains
runtime suspended and also no action taken.
However, when the card in system suspend was not runtime suspended and
the suspend sequence was executed as part of the system suspend
sequence, on system resume there is resume sequence to execute. For
command queueing mode, before suspending the card the CQE was halted,
therefore after resuming the card during system resume the CQE un-halt
must be executed.

This change executes CQE un-halt sequence only in case the card resume
sequence was done.

Change-Id: I41eaa9fe050bbba22cf4a9e118268579f08a1e3c
Signed-off-by: Konstantin Dorfman <kdorfman@codeaurora.org>