bluetooth: Add Nitrous driver for BT power management
Previously there was an issue with the UART Rx wakeup path.  On the firing of
host wake IRQ, kernel should wake up and power on the UART driver to start
receiving data.  The IRQ was not picked up properly because the IRQ handler was
fired before the PM Runtime framewake was ready to wake up the UART driver.
As a result, the Rx data was never received on wakeup and the Nitrous driver
UART tracking state also got out of sync.  This could lead to a wakelock being
held for extended period of time until another Tx or Rx transaction took place.

The driver now detects the firing of host wake IRQ during suspend and delay
the power on of the UART driver until PM resume time.  Polarity configuration
for the host wake and dev wake gpios is also added.

Change-Id: I86654eef4427cc23427075e3e12b41b7782461d7