qcacld-2.0: Fix error check for insufficient buffer in AP stats collection
Replace 'scnprintf' with 'snprintf' to know the number of bytes that were
attempted to be copied while calculating the AP stats. This is needed to
know if the supplied buffer was long enough or not. Since âscnprintfâ
returns only the number of characters written into the buffer, it is not
helpful here and hence replacing it with âsnprintfâ which returns the
number of bytes attempted to copy. snprintf's return value doesn't
include the terminating null byte.

Change-Id: I6364f2d16d96967f31389413f48ccccc66abdefa
CRs-Fixed: 997777
Signed-off-by: Subhani Shaik <subhanis@codeaurora.org>