msm_serial_hs: Move HS UART pins to known state at probe
During probe the driver does a devm_pinctrl_get() on the GPIOs assigned
to it. This call doesn't move the state of pins (bias/drv strength) to a
known state, rather only sets the function mux for those pins. As a result
the pins are in an unknown/floating state till a usecase for the port
begins.
Instead move the pins to the "sleep" state during probe to put them into a
known state.

Change-Id: Ifc6102d367fb0ca9fffbd50a6077dccd779e9a6c
Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>