msm: qpnp-power-on: Hold wakeup event for power key event
On targets with autosleep enabled, it is necessary to hold
wakeup event so that functional units process the event without
interruption. Most devices with wakeup sources has to communicate
the event to userland which does further processing. During this
period system should not attempt autosleep.

Some functional unit detects and processes the wakeup event on their own.
In such cases pm_stay_awake()/pm_relax() can be used to prevent autosleep
during that processing period.

However, for some cases when start and end of processing period is not
defined, use pm_wakeup_event with anticipated timeout. In this case
there are different functional unit in userland which listens to power
key event. Hold a pm_wakeup_event with 2sec of anticipated processing time.

Change-Id: Ie1825b14d2f2917bb1a61175f8886a45dea02bfb
Signed-off-by: Girish S Ghongdemath <girishsg@codeaurora.org>
[Resolved context conflicts]