power: Remove FS sync on suspend
This was detected to introduce races
with Qualcomm's sleepstate driver when
an FS syncs take too long and leads
to the kernel staying awake in some
cases.

Additionally syncing fs on suspend
is completely unnecessary.

This change removes this feature
and adds "DEBUG_FS_SYNC_ON_SUSPEND"
which can be defined for debugging
purposes to reenable FS sync
on suspend.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>

Conflicts:
	kernel/power/suspend.c

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>