LICENSES: Clearly mark dual license only licenses
Just like the CDDL the Apache license and the MPL must only be used as
a choice in additional to an GPL2 compatible license.  Copy over the
boilerplate from the CDDL file to the other two after fixing it up to
make it clear the licenses need to be GPL2 compatible, not just the
more generic GPL compatible.  For example the Apache 2 license is GPL3
compatible, but that doesn't matter for the kernel.

Also move these licenses to a separate directory and document the rules
in license-rules.rst.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Jonathan Corbet <corbet@lwn.net>

Conflict:
   Documentation/process/license-rules.rst